[MS-1229] move hooks into separate files
with index.ts we the keep old import path the same
